requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
PySimpleGUI==4.60.4
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
convertapi-1.7.0
